fix(phase1): Wave 12.9 review round -- a wrapping filter, PHP trim, a vacuous test, two stale docs
Five findings on `5ebf14e`, three P1. Merges Wave 12.6.6's own review fixes in
first, since this branch is stacked on it.

**`month`/`year` wrapped on narrowing (P2).** The filter checked positivity as a
`long` and then narrowed to `int`, so `month=4294967302` became 6 and would have
exported June's batch, where PHP's 64-bit integer stays 4294967302 and matches
nothing. `ExportFilter` carries `Long` through to the bind.
`anOutOfRangeMonthDoesNotWrapIntoARealMonth` asserts both halves -- month 6 finds
the fixture batch, the wrapping value does not -- and fails when the narrowing is
reintroduced.

**The search filter used Java trim (P2).** Java strips every character at or
below U+0020; PHP's `trim()` strips only `" \\t\\n\\r\\0\\x0B"`. A form-feed search
became null here and exported every in-scope payslip, where PHP keeps it and
applies a `LIKE` that matches nothing -- the widest possible divergence from the
narrowest possible input. Now `LegacyValues.phpTrim`, with a regression that
fails under `String.trim()`.

**A vacuous test survived a rename (P1).** `theRemainingUnmappedRoutesAreOnlyTheTwoExports`
kept documentation saying both exports were unmapped while its assertion had been
reduced to checking that delivered routes are present -- it validated no unmapped
condition at all and contradicted the claim beside it. Deleted rather than
rewritten; `allThreeOfC9sOnceUnmappedEndpointsAreNowDelivered` already covers it.

**Two documents contradicted the code (P1).** The endpoint inventory still called
`payslips/export.php` a CSV export with no record of its response contract; it now
carries the XLSX content type, disposition, filename rule, filter set including
the period *overlap* pair, and both column traps. And the completion plan's §8.1
still said all three endpoints were owed and Wave 12.9 was 15 of 16, while §1.6
and §3.2 in the same document said they had shipped -- the section that owns
current sequencing was the stale one. Reconciled, including the note that G2's
numerator closing is not G2 closing.

Validation: `com.workin.legacy.payroll.*` and `LegacyPhpRouteInventoryTest` green;
`python3 scripts/validate_phase0.py` passed; `npx markdownlint-cli2` clean.

+++ backend/src/main/java/com/workin/legacy/payroll/LegacyPayslipStore.java
@@ -54,7 +54,7 @@ public class LegacyPayslipStore {
 	 */
 	public record ExportFilter(
 			long companyId, Long selfEmployeeId, Long employeeId, Long branchId, Long departmentId,
-			Long batchId, Integer month, Integer year, String from, String to, String search) {
+			Long batchId, Long month, Long year, String from, String to, String search) {
 	}
 
 	/**
